xs-controller/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
python-dotenv>=1.0.1
sqlmodel>=0.0.22
aiosqlite>=0.20.0
asyncio-mqtt>=0.16.2
paho-mqtt>=2.1.0
pyjwt>=2.9.0
